backend/ingestion/infra_poller/requirements.txt
requests
redis
psycopg2-binary
